requirements.txt
torch
llama-cpp-python
letta
openclaw
psutil
sounddevice
webrtcvad
fastapi
uvicorn
python-dotenv
pyzmq
